Big theming change
Switched to Gruvbox Material for the dark theme. In order to match it to
tmux better I made significant changes there, however this also meant I
had to have a dark and light mode theme for tmux. So there are also
updates to the `dm` rust tool which can now provide a is_dark flag for
tmux to check using -c , and also forces a reload of tmux config when
switching themes.